#352 New partial for deal show
