+++ resources/js/pages/Deals/Show.vue
@@ -5,6 +5,7 @@ import { ref, onMounted } from 'vue'
 import { type BreadcrumbItem } from '@/types'
 import { route } from 'ziggy-js'
 import { fetchDeal, deleteDeals } from '@/services/dealService'
+import DealDetailSection from './components/DealDetailSection.vue'
 
 interface UserPermissions {
     view: boolean
@@ -37,15 +38,6 @@ const statusClasses: Record<string, string> = {
     won: 'bg-green-100 text-green-700',
     lost: 'bg-red-100 text-red-700',
     archived: 'bg-gray-100 text-gray-600',
-}
-
-function formatCurrency(value: number, currency: string) {
-    return new Intl.NumberFormat('en-GB', { style: 'currency', currency }).format(value)
-}
-
-function formatDate(date?: string | null) {
-    if (!date) return '—'
-    return new Date(date).toLocaleDateString('en-GB')
 }
 
 const props = defineProps<{ deal: any }>()
@@ -133,64 +125,7 @@ onMounted(() => loadDeal())
                     </div>
                 </div>
 
-                <!-- Details -->
-                <div class="grid grid-cols-1 md:grid-cols-2 gap-x-8 gap-y-3 mb-6">
-                    <div>
-                        <span class="font-semibold">Value: </span>
-                        <span>{{ formatCurrency(deal.value, deal.currency) }}</span>
-                    </div>
-                    <div>
-                        <span class="font-semibold">Close Date: </span>
-                        <span>{{ formatDate(deal.close_date) }}</span>
-                    </div>
-                    <div v-if="deal.company">
-                        <span class="font-semibold">Company: </span>
-                        <Link
-                            :href="route('companies.show', { company: deal.company.id })"
-                        >
-                            {{ deal.company.name }}
-                        </Link>
-                    </div>
-                    <div v-if="deal.owner">
-                        <span class="font-semibold">Owner: </span>
-                        <span>{{ deal.owner.name }}</span>
-                    </div>
-                    <div v-if="deal.pipeline">
-                        <span class="font-semibold">Pipeline: </span>
-                        <span>{{ deal.pipeline.name }}</span>
-                    </div>
-                    <div v-if="deal.stage">
-                        <span class="font-semibold">Stage: </span>
-                        <span>{{ deal.stage.name }}</span>
-                    </div>
-                    <div v-if="deal.creator">
-                        <span class="font-semibold">Created By: </span>
-                        <span>{{ deal.creator.name }}</span>
-                    </div>
-                </div>
-
-                <!-- Products -->
-                <div v-if="deal.products && deal.products.length > 0">
-                    <h2 class="text-lg font-semibold border-b pb-2 mb-3">Products</h2>
-                    <table class="w-full border text-sm">
-                        <thead>
-                            <tr>
-                                <th class="p-2 text-left">Product</th>
-                                <th class="p-2 text-right">Qty</th>
-                                <th class="p-2 text-right">Price</th>
-                                <th class="p-2 text-right">Total</th>
-                            </tr>
-                        </thead>
-                        <tbody>
-                            <tr v-for="product in deal.products" :key="product.id" class="border-t">
-                                <td class="p-2">{{ product.name }}</td>
-                                <td class="p-2 text-right">{{ product.pivot?.quantity ?? 1 }}</td>
-                                <td class="p-2 text-right">{{ formatCurrency(product.pivot?.price ?? 0, deal.currency) }}</td>
-                                <td class="p-2 text-right">{{ formatCurrency(product.pivot?.total ?? 0, deal.currency) }}</td>
-                            </tr>
-                        </tbody>
-                    </table>
-                </div>
+                <DealDetailSection :deal="deal" />
 
             </div>
         </div>
